+++ x/erc20/keeper/proposals_test.go
@@ -16,15 +16,17 @@ import (
 const (
 	erc20Name          = "Coin Token"
 	erc20Symbol        = "CTKN"
+	erc20Decimals      = uint8(18)
 	cosmosTokenBase    = "acoin"
 	cosmosTokenDisplay = "coin"
+	cosmosDecimals     = uint8(6)
 	defaultExponent    = uint32(18)
 	zeroExponent       = uint32(0)
 )
 
 func (suite *KeeperTestSuite) setupRegisterERC20Pair() common.Address {
 	suite.SetupTest()
-	contractAddr := suite.DeployContract(erc20Name, erc20Symbol)
+	contractAddr := suite.DeployContract(erc20Name, erc20Symbol, erc20Decimals)
 	suite.Commit()
 	_, err := suite.app.Erc20Keeper.RegisterERC20(suite.ctx, contractAddr)
 	suite.Require().NoError(err)
@@ -233,7 +235,7 @@ func (suite KeeperTestSuite) TestRegisterERC20() {
 		suite.Run(fmt.Sprintf("Case %s", tc.name), func() {
 			suite.SetupTest() // reset
 
-			contractAddr = suite.DeployContract(erc20Name, erc20Symbol)
+			contractAddr = suite.DeployContract(erc20Name, erc20Symbol, cosmosDecimals)
 			suite.Commit()
 			coinName := types.CreateDenom(contractAddr.String())
 			pair = types.NewTokenPair(contractAddr, coinName, true, types.OWNER_EXTERNAL)
@@ -255,8 +257,8 @@ func (suite KeeperTestSuite) TestRegisterERC20() {
 				suite.Require().Equal(coinName, metadata.DenomUnits[0].Denom)
 				suite.Require().Equal(uint32(zeroExponent), metadata.DenomUnits[0].Exponent)
 				suite.Require().Equal(types.SanitizeERC20Name(erc20Name), metadata.DenomUnits[1].Denom)
-				// Default exponent at contract creation is 18
-				suite.Require().Equal(metadata.DenomUnits[1].Exponent, uint32(defaultExponent))
+				// Custom exponent at contract creation matches coin with token
+				suite.Require().Equal(metadata.DenomUnits[1].Exponent, uint32(cosmosDecimals))
 			} else {
 				suite.Require().Error(err, tc.name)
 			}
@@ -280,7 +282,7 @@ func (suite KeeperTestSuite) TestToggleRelay() {
 		{
 			"token not registered",
 			func() {
-				contractAddr = suite.DeployContract(erc20Name, erc20Symbol)
+				contractAddr = suite.DeployContract(erc20Name, erc20Symbol, erc20Decimals)
 				suite.Commit()
 				pair = types.NewTokenPair(contractAddr, cosmosTokenBase, true, types.OWNER_MODULE)
 			},
@@ -290,7 +292,7 @@ func (suite KeeperTestSuite) TestToggleRelay() {
 		{
 			"token not registered - pair not found",
 			func() {
-				contractAddr = suite.DeployContract(erc20Name, erc20Symbol)
+				contractAddr = suite.DeployContract(erc20Name, erc20Symbol, erc20Decimals)
 				suite.Commit()
 				pair = types.NewTokenPair(contractAddr, cosmosTokenBase, true, types.OWNER_MODULE)
 				suite.app.Erc20Keeper.SetERC20Map(suite.ctx, common.HexToAddress(pair.Erc20Address), pair.GetID())
@@ -360,7 +362,7 @@ func (suite KeeperTestSuite) TestUpdateTokenPairERC20() {
 		{
 			"token not registered",
 			func() {
-				contractAddr = suite.DeployContract(erc20Name, erc20Symbol)
+				contractAddr = suite.DeployContract(erc20Name, erc20Symbol, erc20Decimals)
 				suite.Commit()
 				pair = types.NewTokenPair(contractAddr, cosmosTokenBase, true, types.OWNER_MODULE)
 			},
@@ -369,7 +371,7 @@ func (suite KeeperTestSuite) TestUpdateTokenPairERC20() {
 		{
 			"token not registered - pair not found",
 			func() {
-				contractAddr = suite.DeployContract(erc20Name, erc20Symbol)
+				contractAddr = suite.DeployContract(erc20Name, erc20Symbol, erc20Decimals)
 				suite.Commit()
 				pair = types.NewTokenPair(contractAddr, cosmosTokenBase, true, types.OWNER_MODULE)
 
@@ -380,7 +382,7 @@ func (suite KeeperTestSuite) TestUpdateTokenPairERC20() {
 		{
 			"token not registered - Metadata not found",
 			func() {
-				contractAddr = suite.DeployContract(erc20Name, erc20Symbol)
+				contractAddr = suite.DeployContract(erc20Name, erc20Symbol, erc20Decimals)
 				suite.Commit()
 				pair = types.NewTokenPair(contractAddr, cosmosTokenBase, true, types.OWNER_MODULE)
 
@@ -409,8 +411,8 @@ func (suite KeeperTestSuite) TestUpdateTokenPairERC20() {
 				suite.app.BankKeeper.SetDenomMetaData(suite.ctx, banktypes.Metadata{Base: pair.Denom})
 				suite.Commit()
 
-				// Deploy a new contrat with the same values
-				newContractAddr = suite.DeployContract(erc20Name, erc20Symbol)
+				// Deploy a new contract with the same values
+				newContractAddr = suite.DeployContract(erc20Name, erc20Symbol, erc20Decimals)
 			},
 			false,
 		},
@@ -426,8 +428,8 @@ func (suite KeeperTestSuite) TestUpdateTokenPairERC20() {
 				suite.app.BankKeeper.SetDenomMetaData(suite.ctx, metadata)
 				suite.Commit()
 
-				// Deploy a new contrat with the same values
-				newContractAddr = suite.DeployContract(erc20Name, erc20Symbol)
+				// Deploy a new contract with the same values
+				newContractAddr = suite.DeployContract(erc20Name, erc20Symbol, erc20Decimals)
 			},
 			false,
 		},
@@ -443,8 +445,8 @@ func (suite KeeperTestSuite) TestUpdateTokenPairERC20() {
 				suite.app.BankKeeper.SetDenomMetaData(suite.ctx, metadata)
 				suite.Commit()
 
-				// Deploy a new contrat with the same values
-				newContractAddr = suite.DeployContract(erc20Name, erc20Symbol)
+				// Deploy a new contract with the same values
+				newContractAddr = suite.DeployContract(erc20Name, erc20Symbol, erc20Decimals)
 			},
 			false,
 		},
@@ -460,8 +462,8 @@ func (suite KeeperTestSuite) TestUpdateTokenPairERC20() {
 				suite.app.BankKeeper.SetDenomMetaData(suite.ctx, metadata)
 				suite.Commit()
 
-				// Deploy a new contrat with the same values
-				newContractAddr = suite.DeployContract(erc20Name, erc20Symbol)
+				// Deploy a new contract with the same values
+				newContractAddr = suite.DeployContract(erc20Name, erc20Symbol, erc20Decimals)
 			},
 			false,
 		},
@@ -477,8 +479,8 @@ func (suite KeeperTestSuite) TestUpdateTokenPairERC20() {
 				suite.app.BankKeeper.SetDenomMetaData(suite.ctx, metadata)
 				suite.Commit()
 
-				// Deploy a new contrat with the same values
-				newContractAddr = suite.DeployContract(erc20Name, erc20Symbol)
+				// Deploy a new contract with the same values
+				newContractAddr = suite.DeployContract(erc20Name, erc20Symbol, erc20Decimals)
 			},
 			true,
 		},